requirements.txt
SpeechRecognition
gTTS
imageio
opencv-python
ffmpy
moviepy
pygame
pyglet
Pillow
pynput
numpy